+++ src/AccessMessageMetadata/CMakeLists.txt
@@ -9,6 +9,6 @@ add_subdirectory(hs00)
 add_subdirectory(tdct)
 add_subdirectory(ns10)
 add_subdirectory(ep00)
-add_subdirectory(pvCn)
-add_subdirectory(pvAl)
-add_subdirectory(scal)
+add_subdirectory(ep01)
+add_subdirectory(al00)
+add_subdirectory(f144)
