+++ src/AccessMessageMetadata/al00/CMakeLists.txt
@@ -1,0 +1,9 @@
+set(al00_SRC
+        al00_Extractor.cpp
+        )
+
+set(al00_INC
+        al00_Extractor.h
+        )
+
+create_fb_extractor(al00)
