+++ src/AccessMessageMetadata/ep01/CMakeLists.txt
@@ -1,0 +1,9 @@
+set(ep01_SRC
+        ep01_Extractor.cpp
+        )
+
+set(ep01_INC
+        ep01_Extractor.h
+        )
+
+create_fb_extractor(ep01)
