--- src/AccessMessageMetadata/pvAl/CMakeLists.txt
@@ -1,9 +1,0 @@
-set(pvAl_SRC
-        pvAl_Extractor.cpp
-        )
-
-set(pvAl_INC
-        pvAl_Extractor.h
-        )
-
-create_fb_extractor(pvAl)
