--- src/AccessMessageMetadata/pvCn/CMakeLists.txt
@@ -1,9 +1,0 @@
-set(pvCn_SRC
-        pvCn_Extractor.cpp
-        )
-
-set(pvCn_INC
-        pvCn_Extractor.h
-        )
-
-create_fb_extractor(pvCn)
