+++ src/WriterModule/CMakeLists.txt
@@ -10,6 +10,6 @@ add_subdirectory(tdct)
 add_subdirectory(ns10)
 add_subdirectory(f142_test)
 add_subdirectory(ep00)
-add_subdirectory(scal)
-add_subdirectory(pvAl)
-add_subdirectory(pvCn)
+add_subdirectory(f144)
+add_subdirectory(al00)
+add_subdirectory(ep01)
