+++ src/WriterModule/al00/CMakeLists.txt
@@ -1,0 +1,9 @@
+set(al00_SRC
+  al00_Writer.cpp
+)
+
+set(al00_INC
+  al00_Writer.h
+    )
+
+create_writer_module(al00)
