+++ src/WriterModule/al00/al00_Writer.cpp
@@ -7,7 +7,7 @@
 //
 // Screaming Udder!                              https://esss.se
 
-#include "pvAl_Writer.h"
+#include "al00_Writer.h"
 #include "MetaData/HDF5DataWriter.h"
 #include "WriterRegistrar.h"
 #include "json.h"
