+++ src/WriterModule/ep01/CMakeLists.txt
@@ -1,0 +1,10 @@
+set(ep01_SRC
+        ep01_Writer.cpp
+)
+
+set(ep01_INC
+        ep01_Writer.h
+        ConnectionStatusDatasets.h
+)
+
+create_writer_module(ep01)
