+++ src/WriterModule/ep01/ep01_Writer.cpp
@@ -1,4 +1,4 @@
-#include "pvCn_Writer.h"
+#include "ep01_Writer.h"
 #include "FlatbufferMessage.h"
 #include "WriterRegistrar.h"
 #include <ep01_epics_connection_generated.h>
