--- src/WriterModule/pvAl/CMakeLists.txt
@@ -1,9 +1,0 @@
-set(pvAl_SRC
-  pvAl_Writer.cpp
-)
-
-set(pvAl_INC
-  pvAl_Writer.h
-    )
-
-create_writer_module(pvAl)
